WW-5343 Move JUnit4 test case into Struts-core
