+++ plugins/junit/src/main/java/org/apache/struts2/junit/XWorkTestCase.java
@@ -1,0 +1,4 @@
+package org.apache.struts2.junit;
+
+public abstract class XWorkTestCase extends com.opensymphony.xwork2.XWorkTestCase {
+}
